backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
